Part 2.17 done

+++ part2/phonebook/src/components/printPerson.js
@@ -5,10 +5,17 @@ const PrintPerson = ({ person, persons, setPersons }) => {
 
     return (
         <p>{person.name} {person.number} &nbsp;
-            <button onClick={() =>
-                personService.delete(person.id)
-                .then(setPersons(persons.filter(p => p.id !== person.id)))
-            }>
+            <button onClick={() => {
+                if (window.confirm(`Delete ${person.name}?`)) {
+                    personService.delete(person.id)
+                        .then(setPersons(persons.filter(p => p.id !== person.id)))
+                        .catch(error => {
+                            /*Should not ever be needed, but is fail safe is bugs occur*/
+                            alert(`usser ${person.name} was already deleted from the server`)
+                        }
+                        )
+                }
+            }}>
                 delete
     </button></p>
     )
